Implemented additional exceptions for validator classes

+++ server/movies-mananger/src/main/java/com/wiktorkielar/moviesmananger/service/MoviesService.java
@@ -1,17 +1,18 @@
 package com.wiktorkielar.moviesmananger.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.stereotype.Service;
 
-import com.wiktorkielar.moviesmananger.exception.MovieNotFoundException;
 import com.wiktorkielar.moviesmananger.model.Movie;
 import com.wiktorkielar.moviesmananger.model.SortDirection;
 import com.wiktorkielar.moviesmananger.repository.MoviesRepository;
 import com.wiktorkielar.moviesmananger.validator.IdValidator;
+import com.wiktorkielar.moviesmananger.validator.MovieValidator;
 
 @Service
 public class MoviesService {
@@ -20,25 +21,33 @@ public class MoviesService {
 	private MoviesRepository moviesRepository;
 	
 	@Autowired
+	private MovieValidator movieValidator;
+
+	@Autowired
 	private IdValidator idValidator;
 
 	public List<Movie> getMovies(SortDirection sortDirection) throws Exception {
 		Direction direction = Direction.valueOf(sortDirection.name());
 		Sort sort = new Sort(direction, "rating");
+		
 		return moviesRepository.findAll(sort);
 	}
 
 	public Movie createMovie(Movie movie) throws Exception {
+		movieValidator.validateMovie(movie);
+		
 		return moviesRepository.insert(movie);
 	}
 
-	public void deleteMovie(String id) throws Exception {
+	public boolean deleteMovie(String id) throws Exception {
 		idValidator.validateId(id);
-		try {
+		
+		Optional<Movie> optional = moviesRepository.findById(id);
+		if(optional.isPresent()) {
 			moviesRepository.deleteById(id);
-		} catch (IllegalArgumentException e) {
-			throw new MovieNotFoundException("The movie with a given id was not found in the database");
+			return true;
 		}
+		return false;
 		
 	}
